requirements.txt
python-telegram-bot==20.7
playwright==1.18.0
requests==2.26.0